+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/SettingsScreen.kt
@@ -219,6 +219,9 @@ fun SettingsScreen(appUiState: AppUiState, appViewModel: AppViewModel, showVpnSe
 			onMixnetTuningClick = {
 				navController.navigate(Route.MixnetTuning)
 			},
+			onGeoExclusionClick = {
+				navController.navigate(Route.GeoExclusion)
+			},
 		),
 	)
 }